Fixe l'API `save-last-scores` pour convertir les vieux scores en snapshot
